+++ src/v2/mod.rs
@@ -103,7 +103,7 @@ pub use paperclip_core::v2::schema::{self, Schema};
 pub fn from_reader<R, S>(mut reader: R) -> Result<ResolvableApi<S>, PaperClipError>
 where
     R: Read + Seek,
-    for<'de> S: Deserialize<'de> + Schema + Default,
+    for<'de> S: Deserialize<'de> + Schema,
 {
     let mut buf = [0; 1];
     reader.read_exact(&mut buf)?;